% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticVar.R
\name{analyticVar}
\alias{analyticVar}
\title{A Function to read mouse heteroplasmy data}
\usage{
analyticVar(data, normal = FALSE)
}
\arguments{
\item{nameD}{Either "HB" or "LE".}
}
\description{
This function allows you to read mouse heteroplasmy data from external files.
}
\examples{
analyticVar(data)
}
\keyword{data}
\keyword{heteroplasmy}
