% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticVarKimura.R
\name{analyticVarKimura}
\alias{analyticVarKimura}
\title{Kimura-based calculation of the standard error of the variance}
\usage{
analyticVarKimura(data, p, b)
}
\arguments{
\item{data}{The input data in the form of a vector. NA values are omitted.}

\item{p}{The p parameter of the Kimura distribution. Should be in \code{[0,1]}.}

\item{b}{The b parameter of of the Kimura distribution.  Should be in \code{[0,1]}.}
}
\value{
The analytically derived standard error of the variance of \code{data}.
}
\description{
A Kimura-distribution-based method that calculates analytically the standard error of the variance (see Eq. 11 in \emph{Wonnapinij et al., 2010}).
It uses only the Kimura distribution parameters \emph{p} and \emph{b} (passed as arguments).
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
analyticVar(data)

mouseData=readHeteroplasmyData("HB")
mouseData1 = mouseData[which(!is.na(mouseData[,1])),1]
analyticVar(mouseData1)

# use the package data and load it to variable mouseData
mouseData=mousedataLE
# calculate the standard error of the variance for the LE oocyte sample #3
analyticVarKimura(mouseData[,3])
}
\keyword{Kimura}
\keyword{error}
\keyword{heteroplasmy}
\keyword{standard}
\keyword{variance}
