% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapVar.R
\name{bootstrapVar}
\alias{bootstrapVar}
\title{A Function to read mouse heteroplasmy data}
\usage{
bootstrapVar(
  data,
  nrep = 1000,
  biased = FALSE,
  corrected = FALSE,
  sigmaOpt = FALSE
)
}
\arguments{
\item{nameD}{Either "HB" or "LE".}
}
\description{
This function allows you to read mouse heteroplasmy data from external files.
}
\examples{
bootstrapVar(data)
}
\keyword{data}
\keyword{heteroplasmy}
