% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapVarCor.R
\name{bootstrapVarCor}
\alias{bootstrapVarCor}
\title{The bootstrapVar function with default correction}
\usage{
bootstrapVarCor(data, nrep = 1000)
}
\arguments{
\item{data}{The input data in the form of a dataframe or matrix (which will be transformed into
a dataframe). NA values are omitted.}

\item{nrep}{The number of bootstrap resamples. Default is 1000. The higher the number of
the samples, the better the bootstrap outcome (see \code{\link[graphics]{par}}).}
}
\description{
This function is simply the bootstrapVar with the correction argument being TRUE. it is
provided as a seperate function for usability.
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
bootstrapVarCor(data)

mouseData=readHeteroplasmyData("HB")
mouseData1 = mouseData[which(!is.na(mouseData[,1])),1]
bootstrapVarCor(mouseData1,nrep=10000)
}
\references{
\url{https://en.wikipedia.org/}
}
\seealso{
\code{\link{bootstrapVar}}
}
\author{
Kostas and Iain, \email{us@example.com}
}
\keyword{bootstrapVarCor,}
\keyword{fitted}
