% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_ks.R
\name{estimate_parameters_ks}
\alias{estimate_parameters_ks}
\title{Max likelihood estimation of Kimura parameters that minimize the KS statistic}
\usage{
estimate_parameters_ks(h)
}
\arguments{
\item{h}{A vector containig heteroplasmy measurements. Every observation should be in \code{[0,1]}.}
}
\value{
The maximum likelihood estimates for a fitted Kimura distribution parameters that minimze the KS statistic of a KS test.
}
\description{
Using maximum likelihood to estimate the parameters of a fitted Kimura distribution to the input sample values that minimizes the KS statistic. It is used to showcase that using the KS statistic to support selection needs extra caution.
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
estimate_parameters_ml(data_ex)

mouseData=readHeteroplasmyData("LE")
mouseData1 = mouseData[which(!is.na(mouseData[,1])),1]
estimate_parameters_ks(mouseData1)
}
\references{
\href{http://example.com}{Site or paper}
}
\seealso{
\code{\link{readHeteroplasmyData}}
}
\author{
Kostas and Iain, \email{us@example.com}
}
\keyword{heteroplasmy,maximum}
\keyword{likelihood}
