% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroplasmyShift.R
\name{heteroplasmyShift}
\alias{heteroplasmyShift}
\title{Transformed heteroplasmy shift}
\usage{
heteroplasmyShift(h, h0)
}
\arguments{
\item{h}{The heteroplasmy observation. Can be either a single value or a vector of observations.
Every observation should be in \link{0,1}.}

\item{h0}{The reference heteroplasmy value. Should be in \link{0,1}.}
}
\value{
The Transformed heteroplasmy shift.
}
\description{
A numerical transformation of the heteroplasmy samples in order to work with the
heteroplasmy shifts across diverse samples (e.g., due to time or different tissue samples). This
transformation is used for comparing a heteroplasmy observation \emph{h} to a reference value
\ifelse{html}{\emph{h}\out{<sub>0</sub>}}{\eqn{h_0}}.
It corresponds to the formula:\cr \cr
\ifelse{html}{\eqn{\Delta}\out{h =ln((h(h<sub>0</sub> - 1))/(h<sub>0</sub>(h - 1)))}}{\deqn{\Delta h =  \ln \left( \frac{h (h_0 - 1)}{h_0 (h - 1)}\right)}}
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
heteroplasmyShift(data)

mouseData=readHeteroplasmyData("HB")
mouseData1 = mouseData[which(!is.na(mouseData[,1])),1]
heteroplasmyShift(mouseData1,nrep=10000)
}
\references{
\href{http://example.com}{Site or paper}
}
\seealso{
\code{\link{readHeteroplasmyData}}
}
\author{
Kostas and Iain, \email{us@example.com}
}
\keyword{heteroplasmy,transformation,shift}
