% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackVar.R
\name{jackVar}
\alias{jackVar}
\title{A Function to read mouse heteroplasmy data}
\usage{
jackVar(data)
}
\arguments{
\item{nameD}{Either "HB" or "LE".}
}
\description{
This function allows you to read mouse heteroplasmy data from external files.
}
\examples{
jackVar(data)
}
\keyword{data}
\keyword{heteroplasmy}
