% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura_lrt.R
\name{kimura_lrt}
\alias{kimura_lrt}
\title{Likelihood ratio between two heteroplasmy samples}
\usage{
kimura_lrt(h1, h2, use.h0s = F, h1.h0set = 0, h2.h0set = 0)
}
\arguments{
\item{h1}{The first heteroplasmy sample.}

\item{h2}{The two heteroplasmy sample.}

\item{use.h0s}{Logical parameter. TBD}

\item{h1.h0set}{Logical parameter.TBD}

\item{h2.h0set}{Logical parameter.TBD}
}
\value{
The maximum likelihood for the input data according to the Kimura distribution (using bootstrapping)
}
\description{
A function to perform likelihood ratio test exploring difference in bottleneck size between two heteroplasmy samples
}
\examples{
 X.1 = rnorm(50,0.5,0.1)
 X.2 = rnorm(50,0.5,0.1)
kimura_lrt(X.1,X.2)
}
\keyword{statistic}
\keyword{summary}
