% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_dist.R
\name{ks_dist}
\alias{ks_dist}
\title{Kolmogorov-Smirnov distance function}
\usage{
ks_dist(theta, ecdf)
}
\arguments{
\item{theta}{A vector. with two elements. The two Kimura parameters h0 and b.}

\item{ecdf}{A vector containing 10000 values from the empirical cumulative distribution function of the input heteroplasmy data vector. To be used in the optim function in estimate_parameters_ks.}
}
\value{
The maximum distance from the theoretical Kimura distribution.
}
\description{
A function to calculate the Kolmogorov-Smirnov distance. It is used in estimate_parameters_ks to estimate the parameterr that minimize the distance in the optim function.
}
\examples{
h=rnorm(20,0.5,0.1)
ecdf_h <- (stats::ecdf(h))(seq(0, 1, 1e-04))
ks_dist(c(0.5,0.95),ecdf_h)
}
\keyword{Kolmogorov}
\keyword{Smirnov}
\keyword{distance}
