% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxlik.R
\name{maxlik}
\alias{maxlik}
\title{compute maximum likelihood parameters and confidence intervals for heteroplasmy data}
\usage{
maxlik(h, conf.level = 0.95, h0 = F)
}
\arguments{
\item{h}{The heteroplasmy measurements.}

\item{conf.level}{The preferred confidence interval calculation, Default value is 0,95 (95\%).}

\item{h0}{Logical parameter. A particular h0 value  Default is to treat h0 as a fit parameter}
}
\value{
The maximum likelihood for the input data according to the Kimura distribution
}
\description{
compute maximum likelihood parameters and confidence intervals for a given heteroplasmy set.
We can do this while imposing a specific h0 as an argument or allowing a search over h0 values.
}
\examples{
 X.1 = rnorm(50,0.5,0.1)
maxlik(X.1,0.95)
}
\keyword{joint}
\keyword{kimura}
\keyword{likelihood}
\keyword{log}
\keyword{negative}
