% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mousedataBroz}
\alias{mousedataBroz}
\title{3 organellar heteroplasmy samples}
\format{
A data frame with 32 rows and 3 columns (TO FIX)
}
\source{
\url{http://www.example.info/}
}
\usage{
mousedataBroz
}
\description{
3 distinct organellar heteroplasmy samples from Broz et al. (2022). Heteroplasmy data of the package are in the range \code{[0,1]}.(Add reference!)
}
\keyword{datasets}
