% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mousedataFreyer}
\alias{mousedataFreyer}
\title{mouse PGCs heteroplasmy data}
\format{
A data frame with 111 rows and 18 columns (TO FIX)
}
\source{
\url{http://www.example.info/}
}
\usage{
mousedataFreyer
}
\description{
A dataset containing heteroplasmy values from mouse PGCs taken from Freyer et al.
Each column corresponds to a different specimen. NA values have been
added to make the number of rows equal. Please remove them after loading.
Heteroplasmy data of the package are in the range \code{[0,1]}.
}
\keyword{datasets}
