% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mousedataLE}
\alias{mousedataLE}
\title{LE oocyte heteroplasmy data}
\format{
A data frame with 20 rows and 43 columns (TO FIX)
}
\source{
\url{http://www.example.info/}
}
\usage{
mousedataLE
}
\description{
A dataset containing heteroplasmy values for the HB oocyte mouse lines.
Each column corresponds to a different specimen. NA values have been
added to make the number of rows equal. Please remove them after loading.
Heteroplasmy data of the package are in the range \code{[0,1]}.
}
\keyword{datasets}
