% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStdErrVar.R
\name{plotStdErrVar}
\alias{plotStdErrVar}
\title{A Function to read mouse heteroplasmy data}
\usage{
plotStdErrVar(
  data,
  functions = c("normalApr", "analytic", "bootstrap", "correctedBoot", "jackknife"),
  ...
)
}
\arguments{
\item{nameD}{Either "HB" or "LE".}
}
\description{
This function allows you to read mouse heteroplasmy data from external files.
}
\examples{
plotStdErrVar(data)
}
\keyword{data}
\keyword{heteroplasmy}
