% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_kimura_par.R
\name{test_kimura_par}
\alias{test_kimura_par}
\title{Generalised MC KS test for genetic drift}
\usage{
test_kimura_par(h, p, b, num_MC = 1000, round = FALSE)
}
\arguments{
\item{h}{A vector containig heteroplasmy measurements. Every observation should be in \code{[0,1]}.}

\item{p}{The p parameter of the Kimura distribution. Should be in \code{[0,1]}.}

\item{b}{The b parameter of of the Kimura distribution.  Should be in \code{[0,1]}.}

\item{num_MC}{number of Monte Carlo runs}

\item{round}{a logical argument. True if heteroplasmy fractions are rounded to two significant digits. Be cautious as rounding can impact the calculated p-values. Default value is FALSE}
}
\value{
object of class htest
}
\description{
This function is a generalisation of the test_kimura function from the lbozhilova/kimura package. It corresponds to a Monte Carlo Kolmogorov-Smirnov test to detect genetic drift by examining deviation from a Kimura distribution. (add references)
}
\examples{
data_ex=rnorm(n,0.5,0.1)
  fit = estimate_parameters_ml(data_ex)
  p=fit[1]
  b=fit[2]
test_kimura_par(data_ex,p,b)
}
