% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfun.R
\name{transfun}
\alias{transfun}
\title{A Function to cast real numbers to \code{[0,1]}}
\usage{
transfun(x)
}
\arguments{
\item{x}{a real number to be transformed.}
}
\value{
The transformed cast of the input value to the interval \code{[0,1]}.
}
\description{
A transformation function to cast any real number onto the interval \code{[0,1]}.
Equivalent to the inverse logit transform.
}
\examples{
transfun(5.1)
}
\keyword{logit}
\keyword{reverse}
