% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticVar.R
\name{analyticVar}
\alias{analyticVar}
\title{A Function to read mouse heteroplasmy data}
\usage{
analyticVar(data, normal = FALSE)
}
\arguments{
\item{data}{The input data in the form of a dataframe or matrix (which be transformed into
a dataframe).}

\item{normal}{Parameter that indicates if the normal approximation should be used
instead of the general formula from (Wilks, S. S. (1962).Mathematical Statistics).
Default is FALSE.}
}
\description{
This function allows you to read mouse heteroplasmy data from external files.
}
\examples{
analyticVar(data)
}
\keyword{error}
\keyword{heteroplasmy}
\keyword{of}
\keyword{standard}
\keyword{the}
\keyword{variance,}
