% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticVar.R
\name{analyticVar}
\alias{analyticVar}
\title{Analytic calculation of the standard error of the variance}
\usage{
analyticVar(data, normal = FALSE)
}
\arguments{
\item{data}{The input data in the form of a dataframe or matrix (which will be transformed into
a dataframe). NA values are omitted.}

\item{normal}{Parameter that indicates if the normal approximation should be used
instead of the general formula from (Wilks, S. S. (1962).Mathematical Statistics).
Default is FALSE.}
}
\value{
The analytically derived standard error of the variance of \code{data}.
}
\description{
This function calculates analytically the standard error of the variance.
It's based on the use of h-statistic as an estimator
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
analyticVar(data)

mouseData=readHeteroplasmyData("HB")
mouseData1 = mouseData[which(!is.na(mouseData[,1])),1]
analyticVar(mouseData1)

# use the package data and load it to variable mouseData
mouseData=mousedataLE
# calculate the standard error of the variance for the LE oocyte sample #3
bootstrapVar(mouseData[,3])
}
\keyword{error}
\keyword{h-statistic}
\keyword{heteroplasmy}
\keyword{standard}
\keyword{variance}
