% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticVar.R
\name{analyticVar}
\alias{analyticVar}
\title{Analytic calculation of the standard error of the variance}
\usage{
analyticVar(data, normal = FALSE)
}
\arguments{
\item{data}{The input data in the form of a dataframe or matrix (which be transformed into
a dataframe).}

\item{normal}{Parameter that indicates if the normal approximation should be used
instead of the general formula from (Wilks, S. S. (1962).Mathematical Statistics).
Default is FALSE.}
}
\description{
This function calculates analytically the standard error of the variance.
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
analyticVar(data)

mouseData=readHeteroplasmyData("HB")
mouseData1 = mouseData[which(!is.na(mouseData[,1])),1]
analyticVar(mouseData1)
}
\keyword{error}
\keyword{heteroplasmy}
\keyword{of}
\keyword{standard}
\keyword{the}
\keyword{variance,}
