% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters_ml.R
\name{estimate_parameters_ml}
\alias{estimate_parameters_ml}
\title{Max likelihood estimation for Kimura distribution parameters}
\usage{
estimate_parameters_ml(h)
}
\arguments{
\item{h}{A vector containig heteroplasmy measurements. Every observation should be in \code{[0,1]}.}
}
\value{
The maximum likelihood estimates for a fitted Kimura distribution parameters.
}
\description{
Using maximum likelihood to estimate the parameters of a fitted Kimura distribution to the input sample values.
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
estimate_parameters_ml(data_ex)

mouseData=readHeteroplasmyData("LE")
mouseData1 = mouseData[which(!is.na(mouseData[,1])),1]
estimate_parameters_ml(mouseData1)
}
\references{
\href{http://example.com}{Site or paper}
}
\seealso{
\code{\link{readHeteroplasmyData}}
}
\author{
Kostas and Iain, \email{us@example.com}
}
\keyword{heteroplasmy,maximum}
\keyword{likelihood}
