% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invtransfun.R
\name{invtransfun}
\alias{invtransfun}
\title{The inverse function to transfun}
\usage{
invtransfun(x)
}
\arguments{
\item{x}{an integer number to be transformed.}
}
\value{
The inveresed treansformation of transfun.
}
\description{
A transformation function to inveret the effect of transfun (ie, the inverse logit transform).
}
\examples{
invtransfun(0.71)
}
\keyword{inverse}
\keyword{logit}
