% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStdErrVar.R
\name{plotStdErrVar}
\alias{plotStdErrVar}
\title{An example plotting function}
\usage{
plotStdErrVar(
  data,
  functions = c("normalApr", "analytic", "bootstrap", "correctedBoot", "jackknife"),
  ...
)
}
\arguments{
\item{data}{The input data in the form of a dataframe or matrix (which be transformed into
a dataframe).}

\item{functions}{Choose the subset of the functions you wish use for the calculation and
subsequent plot of the standard error of the variance. You can use one or a combination
of "normalApr", "analytic","bootstrap","correctedBoot", and "jackknife". For now,
it outputs all of the aforementioned methods!}
}
\description{
This function is used as a toy example on how to represent the data statistics
regarding the variance of the sample. The mean variance and its standard error
are depicted. Note that this is just an illustration to show that the analytic
and the resampling approaches almost match each other.
}
\examples{
# size of the sample
n=50
#generate a random sample of size n from a normal distribution
data_ex=rnorm(n,0.5,0.1)
plotStdErrVar(data_ex)
}
\keyword{eror}
\keyword{plot}
\keyword{standard}
