% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHeteroplasmyData.R
\name{readHeteroplasmyData}
\alias{readHeteroplasmyData}
\title{A Function to read mouse heteroplasmy data (not finished!)}
\usage{
readHeteroplasmyData(nameD = "HB")
}
\arguments{
\item{nameD}{Either "HB" or "LE" or "Freyer".}
}
\value{
A dataframe containing mouse heteroplasmy data.
}
\description{
This function allows you to read mouse heteroplasmy data from external files.
Use with caution (for now).
}
\examples{
readHeteroplasmyData(nameD="LE")
}
\keyword{data}
\keyword{heteroplasmy}
